backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
scikit-learn==1.5.2
pydantic==2.9.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
